suppressed warnings

+++ testsuite/test_restartFluid/main.cpp
@@ -79,7 +79,7 @@ std::set<UInt> parseList( const std::string& list )
     {
         return setList;
     }
-    UInt commaPos = 0;
+    size_t commaPos = 0;
     while ( commaPos != std::string::npos )
     {
         commaPos = stringList.find( "," );
